ref normalize url encoding
