+++ relay-general/src/store/normalize.rs
@@ -33,6 +33,7 @@ mod mechanism;
 mod request;
 mod spans;
 mod stacktrace;
+mod url_encoding;
 
 pub mod user_agent;
 
@@ -904,9 +905,6 @@ impl<'a> Processor for NormalizeProcessor<'a> {
         _meta: &mut Meta,
         state: &ProcessingState<'_>,
     ) -> ProcessingResult {
-        // TODO: scrubb request.data by trying to convert it to a JSON object and scrubbing
-        // the "variables" object.
-
         request.process_child_values(self, state)?;
 
         request::normalize_request(request)?;
